+++ carpentersblocks/block/BlockCarpentersBed.java
@@ -79,7 +79,7 @@ public class BlockCarpentersBed extends BlockBase
 	/**
 	 * Cycle forward through bed designs.
 	 */
-	protected boolean onHammerRightClick(TECarpentersBlock TE, EntityPlayer entityPlayer, int side, double hitX, double hitZ)
+	protected boolean onHammerRightClick(TECarpentersBlock TE, EntityPlayer entityPlayer, int side, float hitX, float hitZ)
 	{
 		int data = BlockProperties.getData(TE);
 		int design = BedDesignHandler.getNext(Bed.getDesign(data));
